move `TypeLink` component to `@graphiql/react`
