+++ packages/graphiql-react/src/explorer/components/type-link.tsx
@@ -1,19 +1,14 @@
-/**
- *  Copyright (c) 2021 GraphQL Contributors.
- *
- *  This source code is licensed under the MIT license found in the
- *  LICENSE file in the root directory of this source tree.
- */
-
-import { useExplorerContext } from '@graphiql/react';
 import { GraphQLType, isListType, isNonNullType } from 'graphql';
-import React from 'react';
+
+import { useExplorerContext } from '../context';
+
+import './type-link.css';
 
 type TypeLinkProps = {
   type: GraphQLType;
 };
 
-export default function TypeLink(props: TypeLinkProps) {
+export function TypeLink(props: TypeLinkProps) {
   const { push } = useExplorerContext({ nonNull: true, caller: TypeLink });
 
   if (!props.type) {
@@ -37,7 +32,7 @@ export default function TypeLink(props: TypeLinkProps) {
   }
   return (
     <a
-      className="type-name"
+      className="graphiql-doc-explorer-type-name"
       onClick={event => {
         event.preventDefault();
         push({ name: type.name, def: type });
