+++ packages/graphiql/__mocks__/@graphiql/react.tsx
@@ -42,6 +42,7 @@ export {
   StorageContext,
   StorageContextProvider,
   ToolbarButton,
+  TypeLink,
   UnStyledButton,
   useAutoCompleteLeafs,
   useCopyQuery,
