+++ packages/graphiql/src/components/DocExplorer/SearchResults.tsx
@@ -5,11 +5,14 @@
  *  LICENSE file in the root directory of this source tree.
  */
 
+import {
+  TypeLink,
+  useExplorerContext,
+  useSchemaContext,
+} from '@graphiql/react';
 import React, { ReactNode } from 'react';
 
 import Argument from './Argument';
-import TypeLink from './TypeLink';
-import { useExplorerContext, useSchemaContext } from '@graphiql/react';
 import FieldLink from './FieldLink';
 
 export default function SearchResults() {
